+++ src/globaldata.f90
@@ -83,7 +83,7 @@ module filenames
  character(len=120) :: defaultsfile,limitsfile,unitsfile,coloursfile
  integer, dimension(maxfile) :: nstepsinfile
  character(len=*), parameter :: tagline = &
-  'SPLASH: A visualisation tool for SPH data (c)2004-2024 Daniel Price and contributors'
+  'SPLASH: A visualisation tool for SPH data (c)2004-2025 Daniel Price and contributors'
 
  public