+++ src/read_data_sphNG.f90
@@ -1393,13 +1393,14 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  use mem_allocation, only:alloc
  use system_utils,   only:lenvironment,renvironment,envlist
  use labels,         only:ipmass,irho,ih,ix,ivx,labeltype,print_types,headertags,&
-                          iutherm,itemp,ikappa,irhorestframe,labelreq,nreq
+                          iutherm,itemp,ikappa,irhorestframe,labelreq,nreq,get_sink_type
  use calcquantities, only:calc_quantities
  use asciiutils,     only:make_tags_unique,match_tag
  use sphNGread
  use lightcurve_utils, only:get_temp_from_u,ionisation_fraction,get_opacity
- use read_kepler,      only:check_for_composition_file,read_kepler_composition
+ use readcomposition,  only:check_for_composition_file,read_composition
  use byteswap,         only:bs
+ use part_utils,       only:locate_nth_particle_of_type
  integer, intent(in)  :: indexstart,iposn
  integer, intent(out) :: nstepsread
  character(len=*), intent(in) :: rootname
@@ -1413,7 +1414,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  integer :: ipos,nptmass,nptmassi,ndust,nstar,nunknown,ilastrequired
  integer :: imaxcolumnread,nhydroarraysinfile,nhdr,nkilled
  integer :: itype,iphaseminthistype,iphasemaxthistype,nthistype,iloc,idenscol
- integer :: icentre,icomp_col_start,ncomp,nextracols
+ integer :: icentre,icomp_col_start,ncomp,nextracols,isink1
  integer, dimension(maxparttypes) :: npartoftypei
  real,    dimension(maxparttypes) :: massoftypei
  logical :: iexist, doubleprec,imadepmasscolumn,gotbinary,gotiphase
@@ -1434,7 +1435,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
  real, dimension(maxinblock) :: dummyreal
  real :: hfact,omega
  real(doub_prec) :: Xfrac,Yfrac
- real :: xHIi,xHIIi,xHeIi,xHeIIi,xHeIIIi,nei
+ real :: xHIi,xHIIi,xHeIi,xHeIIi,xHeIIIi,nei,m1,rad_corotate
  logical :: skip_corrupted_block_3,get_temperature,get_kappa,get_kappa_tot
  logical :: get_ionfrac,need_to_allocate_iphase,got_tag
  character(len=lentag) :: tagsreal(maxinblock), tagtmp
@@ -1966,6 +1967,11 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
                 print "(a)",'ERROR: not enough arrays written for sink particles in phantom dump'
                 nskip = nreal(iarr)
              else
+                nskip = nint(iarr)
+                do i=1,nskip
+                   if (tagged) read(iunit,end=33,iostat=ierr) ! skip tags
+                   read(iunit,end=33,iostat=ierr)
+                enddo
                 if (debug) print*,'DEBUG: denoting ',npart,'->',npart+isize(iarr),' as sink particles'
                 iphase(npart+1:npart+isize(iarr)) = -3
                 ilocvx = nreal(iarr)-2 ! velocity is always last 3 numbers for phantom sinks
@@ -2266,7 +2272,7 @@ subroutine read_data_sphNG(rootname,indexstart,iposn,nstepsread)
 endif
 
  if (icomp_col_start > 0 .and. any(required(icomp_col_start:icomp_col_start+ncomp))) then
-    call read_kepler_composition(compfile,ntotal,dat(:,:,j),icomp_col_start,ncomp)
+    call read_composition(compfile,ntotal,dat(:,:,j),icomp_col_start,ncomp)
  endif
  !
  !--calculate the temperature from density and internal energy (using physical units)
@@ -2301,38 +2307,6 @@ endif
        dat(i,iHeIIcol,j)=xHeIIi
        dat(i,iHeIIIcol,j)=xHeIIIi
     enddo
- endif
-
- !
- !--reset centre of mass to zero if environment variable "SSPLASH_RESET_CM" is set,
- !  or reset centre to densest clump if environment variable "SSPLASH_RESET_DENSE" is set
- !  the latter will override the former
- ! (updated from n1 to npart since order is not preserved when dumping data; JHW)
- icentre = 0
- if (lenvironment('SSPLASH_RESET_CM'))    icentre = 1
- if (lenvironment('SSPLASH_RESET_DENSE')) icentre = 2
- if (allocated(dat) .and. npart > 0 .and. npart <= size(dat(:,1,1)) .and. icentre > 0 .and. allocated(iphase)) then
-    call reset_centre_of_mass(dat(1:npart,1:3,j),dat(1:npart,4,j),dat(1:npart,5,j),iphase(1:npart),npart,icentre)
- endif
- !
- !--reset corotating frame velocities if environment variable "SSPLASH_OMEGA" is set
- !
- if (allocated(dat) .and. n1 > 0 .and. all(required(1:2))) then
-    omega = renvironment('SSPLASH_OMEGAT')
-    if (abs(omega) > tiny(omega) .and. ndim >= 2) then
-       call reset_corotating_positions(n1,dat(1:n1,1:2,j),omega,time(j))
-    endif
-
-    if (.not. smalldump) then
-       if (abs(omega) < tiny(omega)) omega = renvironment('SSPLASH_OMEGA')
-       if (abs(omega) > tiny(omega) .and. ivx > 0) then
-          if (.not.all(required(1:2)) .or. .not.all(required(ivx:ivx+1))) then
-             print*,' ERROR subtracting corotating frame with partial data read'
-          else
-             call reset_corotating_velocities(n1,dat(1:n1,1:2,j),dat(1:n1,ivx:ivx+1,j),omega)
-          endif
-       endif
-    endif
  endif
 
  !--set flag to indicate that only part of this file has been read
@@ -2468,12 +2442,6 @@ endif
 
  endif iphasealloc
 
- if (allocated(dattemp)) deallocate(dattemp)
- if (allocated(dattempsingle)) deallocate(dattempsingle)
- if (allocated(dattemp2)) deallocate(dattemp2)
- if (allocated(iphase)) deallocate(iphase)
- if (allocated(listpm)) deallocate(listpm)
-
  call set_labels_sphNG
  if (.not.phantomdump) then
     if (ngas /= npart - nptmassi - ndust - nstar - nunknown) &
@@ -2493,9 +2461,57 @@ endif
     npartoftype(itypemap_unknown_phantom,j) = npartoftype(itypemap_unknown_phantom,j) + nunknown
  endif
 
+
+ !
+ !--reset centre of mass to zero if environment variable "SSPLASH_RESET_CM" is set,
+ !  or reset centre to densest clump if environment variable "SSPLASH_RESET_DENSE" is set
+ !  the latter will override the former
+ ! (updated from n1 to npart since order is not preserved when dumping data; JHW)
+ icentre = 0
+ if (lenvironment('SSPLASH_RESET_CM'))    icentre = 1
+ if (lenvironment('SSPLASH_RESET_DENSE')) icentre = 2
+ if (allocated(dat) .and. npart > 0 .and. npart <= size(dat(:,1,1)) .and. icentre > 0 .and. allocated(iphase)) then
+    call reset_centre_of_mass(dat(1:npart,1:3,j),dat(1:npart,4,j),dat(1:npart,5,j),iphase(1:npart),npart,icentre)
+ endif
+ !
+ !--reset corotating frame velocities if environment variable "SSPLASH_OMEGA" is set
+ !
+ if (allocated(dat) .and. n1 > 0 .and. all(required(1:2))) then
+    rad_corotate = renvironment('SSPLASH_OMEGAR')
+    call locate_nth_particle_of_type(1,isink1,get_sink_type(ntypes),iamtype(:,j),npartoftype(:,j),ntotal)
+    if (rad_corotate > tiny(rad_corotate) .and. isink1 > 0) then
+       m1 = dat(isink1,4,j)
+       omega = -sqrt(m1 / rad_corotate**3)
+       print "(a,g8.2,a,g8.2)",' :: COROTATING FRAME at R = ',rad_corotate,' around sink particle #1, M=',m1
+    else
+       omega = renvironment('SSPLASH_OMEGAT')
+    endif
+    if (abs(omega) > tiny(omega) .and. ndim >= 2) then
+       call reset_corotating_positions(n1,dat(1:n1,1:2,j),omega,time(j))
+    endif
+
+    if (.not. smalldump) then
+       if (abs(omega) < tiny(omega)) omega = renvironment('SSPLASH_OMEGA')
+       if (abs(omega) > tiny(omega) .and. ivx > 0) then
+          if (.not.all(required(1:2)) .or. .not.all(required(ivx:ivx+1))) then
+             print*,' ERROR subtracting corotating frame with partial data read'
+          else
+             call reset_corotating_velocities(n1,dat(1:n1,1:2,j),dat(1:n1,ivx:ivx+1,j),omega)
+          endif
+       endif
+    endif
+ endif
+
  if (iverbose > 0) call print_types(npartoftype(:,j),labeltype)
 
  close(15)
+
+ if (allocated(dattemp)) deallocate(dattemp)
+ if (allocated(dattempsingle)) deallocate(dattempsingle)
+ if (allocated(dattemp2)) deallocate(dattemp2)
+ if (allocated(iphase)) deallocate(iphase)
+ if (allocated(listpm)) deallocate(listpm)
+
  if (debug) print*,' finished data read, npart = ',npart, ntotal, npartoftype(1:ntypes,j)
 
  return
@@ -2763,6 +2779,22 @@ subroutine set_labels_sphNG
           igrainsize = i
        case('graindens')
           igraindens = i
+       case('grainmass')
+          label(i) = 'grain mass'
+          units(i) = umass
+          unitslabel(i) = ' [g]'
+       case('csound')
+          label(i) = 'sound speed'
+          units(i) = udist/utime
+          unitslabel(i) = ' [cm/s]'
+       case('rhogas')
+          label(i) = '\rho_g'
+          units(i) = umass/udist**3
+          unitslabel(i) = ' [g/cm^3]'
+       case('dv')
+          label(i) = 'dv'
+          units(i) = udist/utime
+          unitslabel(i) = ' [cm/s]'
        case('temperature')
           itemp = i
        case('vrel')
@@ -2823,8 +2855,8 @@ subroutine set_labels_sphNG
     enddo
     ! Make N deltav labels with vector subscripts
     do i = ideltavsum+ndimV,ideltav,ndimV
-       write(deltav_string,'(I10)') (i-ideltavsum)/ndimV
-       write(deltav_string,'(A)') 'deltav'//trim(adjustl(deltav_string))
+       write(deltav_string,'(i10)') (i-ideltavsum)/ndimV
+       write(deltav_string,'(a)') 'deltav'//trim(adjustl(deltav_string))
        do j=1,ndimV
           label(i+j-1) = trim(deltav_string)//'_'//labelcoord(j,1)
        enddo